+++ Documentation/Changelog.md
@@ -10,8 +10,9 @@ Release date: not released yet
 - Added support for [`FORCE_COLOR` convention](https://force-color.org/)
 - Added `CommandLine.EncodeFileName(string)` method which can be used to ensure that file names are interpreted as command-line arguments rather than options (e.g., prevents names starting with `-` from being misinterpreted)
 - Added path comparison-sound equality operations for `VfsLocation` and `VfsReadOnlyLocation`
-- Improved nullability annotations
+- Added polyfill for `Marshal.GetLastPInvokeError()` method
 - `AppInformation.Current.ExecutablePath`  can point to a launcher file only if it is proven by actual process information
+- Fixed nullability annotation inconsistencies
 - Fixed parameter name escaping in `UriQueryBuilder.HasParameter(string)` method
 - Fixed handling of `\x1b[m` control sequence in `Gapotchenko.FX.Console` module
 - Fixed issue with `Graph<T>.IsCyclic` calculation for undirected graphs